+++ lib/src/populate.rs
@@ -136,7 +136,7 @@ pub fn populate_base_models(store: &impl Storelike) -> AtomicResult<()> {
             urls::PARENT.into(),
             Value::AtomicUrl("https://atomicdata.dev/properties".into()),
         );
-        store.add_resource_opts(&resource, false, false, true)?;
+        store.add_resource_opts(&resource, false, true, true)?;
     }
 
     for c in classes {
@@ -145,7 +145,7 @@ pub fn populate_base_models(store: &impl Storelike) -> AtomicResult<()> {
             urls::PARENT.into(),
             Value::AtomicUrl("https://atomicdata.dev/classes".into()),
         );
-        store.add_resource_opts(&resource, false, false, true)?;
+        store.add_resource_opts(&resource, false, true, true)?;
     }
 
     Ok(())
@@ -275,23 +275,19 @@ pub fn populate_sidebar_items(store: &crate::Db) -> AtomicResult<()> {
         format!("{}/import", base),
         format!("{}/collections", base),
     ];
-    drive.set_propval(urls::SUBRESOURCES.into(), arr.into(), store)?;
+    for item in arr {
+        drive.push_propval(urls::SUBRESOURCES, item.into(), true)?;
+    }
     drive.save_locally(store)?;
     Ok(())
 }
 
 /// Runs all populate commands. Optionally runs index (blocking), which can be slow!
 #[cfg(feature = "db")]
-pub fn populate_all(store: &crate::Db, index: bool) -> AtomicResult<()> {
+pub fn populate_all(store: &crate::Db) -> AtomicResult<()> {
     // populate_base_models should be run in init, instead of here, since it will result in infinite loops without
     populate_default_store(store)
         .map_err(|e| format!("Failed to populate default store. {}", e))?;
-    // This is a potentially expensive operation, but is needed to make Queries work with the models created in here
-    if index {
-        store
-            .build_index(true)
-            .map_err(|e| format!("Failed to build index. {}", e))?;
-    }
     create_drive(store).map_err(|e| format!("Failed to create drive. {}", e))?;
     set_drive_rights(store, true)?;
     populate_collections(store).map_err(|e| format!("Failed to populate collections. {}", e))?;