+++ server/e2e_tests/e2e-generated.spec.ts
@@ -184,7 +184,7 @@ test.describe('data-browser', async () => {
       '[data-test="sort-https://atomicdata.dev/properties/description"]',
     );
     // These values can change as new Properties are added to atomicdata.dev
-    const firstPageText = 'text=A base64';
+    const firstPageText = "text=A base64 serialized JSON object";
     const secondPageText = 'text=include-nested';
     await expect(page.locator(firstPageText)).toBeVisible();
     await page.click('[data-test="next-page"]');