+++ CHANGELOG.md
@@ -12,6 +12,7 @@ See [STATUS.md](server/STATUS.md) to learn more about which features will remain
 - Add Table Ontology #25
 - Fix Post endpoints not including search params in returned `@id` field.
 - Rebuilding indexes done on separate thread, only once #616 #615
+- Don't require building index for populate commands
 
 ## [v0.34.2] - 2023-03-04